Split out DX11 alloc resources from accel file
